Remove styled-components dependency from regl-worldview
We only had a few minor dependencies on it, but it caused a bunch of
problems, ranging from hacks in the Rollup config, to having to keep
the version in sync with the version in the root of the repo (for docs).
So I figured I’d just remove it.

Test plan: verified manually that stories and docs work properly now.
The rest is somewhat covered by CI.