Update flow-typed definitions (#22)
And make the process for updating those definitions a bit easier. And
get rid of flow-mono since we don’t really need it and it was failing
silently when building flow-typed definitions.

And fixed a definition in WorldviewContext with Audrey’s help.

Also updated the readme.

Test plan: Flow works!