Fix screenshot tests not working properly [EP-4055]
I wasn’t able to get WebGL working in a CircleCI docker container, not
even when pointing it to our custom image for which I included a
Dockerfile.

I was only able to make it work when using `machine: true` and then
running Docker directly. This is slightly slower but still pretty fast,
and gives us most control over how Docker is run.

Test plan: exported the screenshots as CircleCI artifacts so we can look
at them until we have reg-suit or so set up (for which we’re blocked on
getting an S3 bucket).