Add a few more packages to this repo
Just some small packages, in preparation for the bigger addition of
`webviz-core`.

Note that I opted to use `cruise-button` and `cruise-tooltip` instead of
`@cruise/button` and `@cruise/tooltip`. This is to make the setup with
Lerna and Webpack (when we add stories and `webviz-core`) a bit simpler.

Also reran `npm run flow-typed-rebuild` to get some new definitions for
the new dependencies.

Test plan: verified manually that these packages work when copied over
to the node_modules directory of our proprietary repo. Also made sure
tests/flow/lint works.