Add webviz-core (#53)
This adds the main package that we’re open sourcing, a rosbag analysis
tool. For now the webviz-core repo is kept in sync with a directory in
our internal repository. I added some more information to the README.md
about that.

This also means that the monorepo tools have to stay in sync with our
internal tools and versions. This PR does some of that by setting
specific versions, copying in flow-typed definitions from our internal
repo, updating configurations, and so on.

I also had to remove styled-components from `@cruise-automation/button`
as it was conflicting with other styles. I could have figured out how
to deduplicate it but it was easier to just remove that dependency
altogether. :)

This also hosts the Webviz in the root of the repo. To get that to work
while still keeping index.html in the package directory, I used a
symlink. Similarly I added a symlink for the wasm-lz4.wasm dependency.
When releasing using `gh-pages`, we overwrite those symlink files.

Finally, I removed remove preset-env. It adds a bunch of bloat and is not
necessary since we only target modern browsers. Per Brian’s suggestion.

Test plan: tested manually locally; pushed up to
https://cruise-automation.github.io/webviz/ (try it out!), covered by
lint/flow/tests.