Fix reading hitmap on every mouse move
We had some incorrect logic that would cause the Worldview component
to read from the hitmap on every mouse move even if hitmapOnMouseMove
was false, as long as there was no `onMouseMove` handler. Fix it and
make the logic path a bit clearer.

Test plan: test manually with interactions to make sure they aren't
broken.