Add Worldview interactivity integration tests (#212)
## Summary

Add a bunch of integration tests for worldview interactivity. These integration tests take the form of Storybook stories. The assertions run in the browser immediately upon loading the page and, if they fail, display the stack trace instead of the component being rendered. We test these in CI using our screenshot tests, so we will know if one of the tests have failed.

Tests cover the `onClick` handler for each component and the worldview-level `onClick`, and test the most common cases: clicking an object, clicking an object when `enableStackedObjectEvents` is enabled, and instanced clicking if applicable. There are also basic tests for other events.

This testing enabled discovery and fixing of the following bugs, mostly involving errors around stacked object events:
* Sometimes we would complain that the `getChildrenFromHitmap` function was not respecting `excludedObjects` in stacked object hitmaps when it really was
* Ensure that we print errors when we encounter a bitmap color that we don’t know about
* GLTFScene: did not return the exact marker passed into children when evented
* Grid: could not be clicked, incorrect shader code
* `getChildrenForHitmapWithOriginalMarker` would not respect `excludedObjects` in stacked object hitmaps at all
* `createInstancedGetChildrenForHitmap` would not respect `excludedObjects` in stacked object hitmaps in many cases

## Test plan

These are tests. There is also a deliberately "failing" screenshot test and jest tests for our internally built expect matchers.

## Versioning impact

This will require bumping the minor version of the package.