Fix codesandbox links that include common/
We import common/ as an absolute import, which doesn't play nicely
with codesandbox. The documented ways to enabling aliased import
resolution in codesandbox don't seem to work, so instead let's just
find and replace all imports so that they can be imported in a
relative way in codesandbox.

Test plan: tested manually