Add option to only render triangles in the hitmap (#266)
## Summary

For performance reasons we sometimes want to only render triangles when clicking. Add an option to allow this. Modify the command API somewhat to enable this. Also do a small cleanup: move all of the `createInstancedGetChildrenForHitmap` calls out of the render functions of commands.

## Test plan

Added a story with this new behavior.

## Versioning impact

Minor package increment needed.