Rename /try to /app
Since “try” implies that this is just a demo version of the app, when
it’s actually our intended production version. “app” communicates that a
bit better.

I left a page with a redirect that preserves query parameters under
`/try` so that any existing applications will keep working.

Test plan: tried this locally, and did a search of “/try” and “try/“ in
our codebase, which I think should cover all references.