Worldview: fix hitmap bug
I'm seeing an error in our logs: "(regl) can not resize a framebuffer
which is currently in use" in the readHitmap function. Fix by queueing
the readHitmap function because the framebuffer can only be used in
one function at a time.

Test plan: test added for DebouncedPromise but existing interaction tests
should test for this.

Release plan: minor release required