Replace recursive connection logic with a loop
This change introduces an explicit "Keep starting new connections until
we run out of things to do" loop.

The mutual recursion between _updateState and _setConnection was
confusing, and (I guess) could have issues with stack depth if there
were a lot of open read-requests. That could happen if the user scrubbed
around a lot, for example.

This change will also make it easier to introduce explicit logic to
preload data when we have memory headroom.