Move request-resolving to a less confusing place
We want to resolve any completed requests after any block has been
downloaded. We know `this.readRequests[0]` will only be done after this
connection has finished, but other requests might get completed in the
middle of this connection.

Moreover, we should do it before purging, because there's a slightly
greater chance we'll have something to resolve at that point.