MemoryCacheDataProvider prefetches more data for "large" files (#356)
* Pull out _resolveFinishedReadRequests into a method

Currently _setConnection recursively calls _updateState, which is a bit
confusing -- mutual recursion, along with mutable state, along with
asynchronous evaluation... Removing the recursion might make things a
bit easier to reason about.

To that end, _setConnection should exit on completion, be called in a
loop, and need only call _resolveFinishedReadRequests, not call
_updateState.

* Pull out _maybeRunNewConnections into a method

This method will be used to run the "connection loop" that is currently
implicit in the mutually-recursive _setConnection/_updateState logic.

* _setConnection signals success

In replacing the recursive logic with an explicit loop,
the loop needs to know whether

a) The connection was interrupted, in which case it isn't responsible
   for creating new connections, or
b) The connection completed successfully, in which case we should try to
   complete any more pending requests.

To distinguish between those cases, _setConnection returns a bool.

* Replace recursive connection logic with a loop

This change introduces an explicit "Keep starting new connections until
we run out of things to do" loop.

The mutual recursion between _updateState and _setConnection was
confusing, and (I guess) could have issues with stack depth if there
were a lot of open read-requests. That could happen if the user scrubbed
around a lot, for example.

This change will also make it easier to introduce explicit logic to
preload data when we have memory headroom.

* Uniform prefetching between big and small files

Prefetching is now done explicitly, by identifying times when there is
free space in the cache and there are no pending requested ranges, and
creating connections for uncached ranges.

The "minimum cache size" parameter is removed, and "small files" fill up
their caches by the same mechanism as large files.

This behaviour is pretty awkward to test. Complications in deciding
"correct" behaviour include,
 - Prefetching, which doesn't know the byte size of what it's fetching,
 - `READ_AHEAD_BLOCKS`, which can result in more being fetched than is
   asked for, and
 - `getBlocksToKeep` leaving the cache slightly over-full.

* Move request-resolving to a less confusing place

We want to resolve any completed requests after any block has been
downloaded. We know `this.readRequests[0]` will only be done after this
connection has finished, but other requests might get completed in the
middle of this connection.

Moreover, we should do it before purging, because there's a slightly
greater chance we'll have something to resolve at that point.