Evict cached blocks from end of ranges sometimes
Existing bug: Seeking to a soon-to-be-evicted block during playback
results in an infinite loop in which a block is requested, immediately
evicted, and then requested again etc.

This change fixes the bug by trying to not evict blocks in those
circumstances: When there are open read requests, we avoid evicting from
the edge of blocks closest to the first ("highest priority") request.

If there are no open read requests, perhaps because we've just resolved
the last one, try not to evict blocks near the last resolved request too
-- odds are fair we'll request another block nearby soon.