Stop NodeDataWorker in UserNodePlayer on close
We're currently leaking a 50MB `NodeDataWorker` every time we open a
file. (Most of the weight is probably typescript?) We should explicitly
terminate the worker instead.

Alternatively, we could reuse these -- either by keeping references to
old unused workers, or by making them static and never terminating them
(assuming they're stateless.) To inform this decision, I've measured the
time to instantiate a `NodeDataWorker` in an `Rpc` and send it a dummy
message. It takes between 1.5 and 1.6 seconds on my personal laptop.

A design question, answerable later: Should `Rpc` have a `terminate()`
method? I imagine this sort of thing is pretty common.

Also something to fix later: On application startup, we also leak a
650KB roslib "workersocket" every second. Not much, but it's over 2 gigs
per hour so it could occasionally cause users pain.