Update docker CI image
This is a fix for a long-winded attempt to get packages synced between
internal and open source versions. We use NPM aliasing internally, but
it doesn't seem to work in open-source - see #362, which keeps failing
the build even though the build works locally. It turns out that the
version of NPM in CI (6.4.2) doesn't support package aliasing while
my local version of NPM does - it turns out that support was only
added in 6.9.0 (https://npm.community/t/release-npm-6-9-0/5911).

The fix is simple: update the dockerfile version that we use to
be able to use package aliasing in CI.