Fix build memory usage (#429)
Because even though we bumped the limit in the Dockerfile, the Docker
Hub still only gives us an instance with 2GB RAM (see
https://success.docker.com/article/what-are-the-current-resource-limits-placed-on-automated-builds).

Test plan: Couldn’t actually test if this fixes the automated build though, because
I ran into this issue: https://forums.docker.com/t/422-unprocessable-entity-when-configuring-the-automated-build-in-docker-hub/70527/2
when trying to add a different branch for autobuilding. :( But it works
for me locally without the limit increase in the Dockerfile, so that’s
encouraging.