Optimize Dockerfile for smaller container images
This change uses a multi-stage build to install build-time dependencies into
a temporary container, while only the build products are copied into the final
container image, a standard nginx container.

The Dockerfile is now structured to take advantage of the Docker cache when
re-building. Code changes that do not alter package.json etc. will not require
a complete re-installation of dependencies.