Less buffer copying in <Lines>
Whenever we render a line we copy over the whole buffer, which is as
long as the longest line we've ever seen. This can be very inefficient
if there are a small number of very long lines and a large number of
short lines.

Test plan: No functional changes. Screenshots should stay unchanged.