Switch withTransforms back to a HoClass component
React is complaining that returning a function component here breaks the "rules of hooks".
I messed with it for a while trying to get it to work, but couldn't. It might be the double layer of `hoistNonReactStatics`/`React.forwardRef` in both `FrameCompatibilityDEPRECATED` and `withTransforms` that's causing the problem.
I'm not entirely sure, but switching it back to a class component fixes the problem.