Worldview: Paint again if onDirty was called during paint
In some cases, since components can register arbitrary functions to be called during paint(), painting may trigger a React state update, which causes us to call onDirty again. But since a paint is currently in progress, we just ignore the onDirty call, which means that the new state never gets drawn until another paint is triggered in the future.

Fix this by keeping track of whether onDirty is called during paint, and if so just queue up another frame.